chore: Refactor module imports and update dependencies
